Removed all redundant / already created code from us merging our branches together!
