--- alice_public.pem
@@ -1,14 +1,0 @@
------BEGIN PUBLIC KEY-----
-MIICIjANBgkqhkiG9w0BAQEFAAOCAg8AMIICCgKCAgEA3Z08V+5oIo1guY1nwDm9
-CJuqivhhcINCwz0Zn4rh8RflNpzoFOfmw3VO2NMdujhOi1cdb7VYBN2xPAbbEeMV
-PYjkFZqdL+dCdvWPtvAcW1W2OXonUogmxY63q5lmq/6tvBukgA9HDdjzxYHV/aLd
-LhzptjoMnLgwEGXt+Wzkn+qkMjtNOxbRDlYCtgjaRdezVV8ZEfuzcz8BhzcvV4i5
-Nxv7QoBEWQoU3cfhL5TYR9vBfmoAScyKTtcTRHQjnH+T22KIsiYDxXuZRYcS/9ir
-cR7AdC8zDCivODzV+KsZ54PxvYlNQCfDYzeucaiyQHADNPb2PtLzQpqeGnFtr49i
-oF2aazkgz047zR1R8ftNUeihNckML2wu6fgIf8ckx+T+yF3peOWi2Xa1ci5S7+nD
-BJ6HIs7RkwWT7km0OgJL3vh7xMcfQw8FnTA6biFgoDDmFgGpSq7gLTGuVRRw3qsq
-+oFEX/Mj95FEt/sEGXBxtQ5b8Xcs9DliczVoa0CLKKOka3qr2SO8xEvo/PtGPW/U
-7W5v66VSRCAdV4JIAovkBcLV3+ARxY9mEfPm8w+tu/Hk633yHqBdP+XBfChaEQ2A
-rodsdiOXovAuhesANNhS61b2EAUgdHaGdLzp/KXJpkiO99F1+8NpzXtNWh7KpyHv
-I4dabBykTcPf0tow6F94hnUCAwEAAQ==
------END PUBLIC KEY-----
