--- protocol/types.py
@@ -1,18 +1,0 @@
-from __future__ import annotations
-
-#client->server
-DIR_GET_PUBKEY = "DIR_GET_PUBKEY"
-DIR_GET_WRAPPED_PUBLIC_KEY = "DIR_GET_WRAPPED_PUBLIC_KEY"
-
-#server->client
-DIR_PUBKEY = "DIR_PUBKEY"
-DIR_WRAPPED_PUBLIC_KEY = "DIR_WRAPPED_PUBLIC_KEY"
-ERROR = "ERROR"
-
-# Error Codes
-ERR_BAD_JSON = "BAD_JSON"
-ERR_UNKNOWN_TYPE = "UNKNOWN_TYPE"
-ERR_USER_NOT_FOUND = "USER_NOT_FOUND"
-ERR_NO_WRAPPED_KEY = "NO_WRAPPED_KEY"
-ERR_DIR_LOOKUP_FAILED = "DIR_LOOKUP_FAILED" 
-ERR_NAME_IN_USE = "NAME_IN_USE"
