--- server/main.py
@@ -1,2 +1,0 @@
-import uuid
-print(str(uuid.uuid4()))
